requirements_tf_1x.txt
efficientnet==1.0.0
image-classifiers==1.0.0
Keras==2.3.1
numpy==1.19.4
nvsmi==0.4.2
pandas==1.2.4
tensorflow-gpu==1.15.0
tqdm==4.46.0
